refactor(api): split details.js controller into per-entity modules (#599)
details.js had grown to 915 lines (audit finding C1). Split it by
responsibility into focused modules under routes/details/:

  user.js          — /api/user/:id + sub-resources
  group.js         — /api/group/:id + sub-resources
  accessPackage.js — /api/access-package/:id + sub-resources
  shared.js        — cleanRow / getPermissionTable / fetchHistory /
                     countHistory + useSql / UUID_RE constants

details.js is now a 26-line barrel that composes the three routers, so
app.js and every public path are unchanged. Pure code move — the only
behavioural-adjacent change is the categories.js dynamic-import path
(./categories.js -> ../categories.js) now that accessPackage.js lives a
directory deeper.

The two/three handlers that were grandfathered over the complexity
threshold keep their exact values; their baseline entries are relocated
from details.js to the new file paths (no increase), matching the
matrix.js split precedent.

Added shared.test.js covering the extracted helpers, which raises the
routes/details branch coverage. Full API suite green (1706 tests),
coverage above every floor, both complexity gates and the OpenAPI-drift
guard pass.
